+++ test/fixtures/another_dependency.rb
@@ -1,2 +1,3 @@
+require 'opal'
 require 'dependency'
 require 'inside_load_path'