+++ test/fixtures/js_require.rb
@@ -1,2 +1,3 @@
+require 'opal'
 require 'pure_js'
 require 'pure_js2.js'