+++ test/fixtures/relative.rb
@@ -1,2 +1,3 @@
+require 'opal'
 # sprockets allows you to not lead with a '.'
 require_relative 'tree/file1'