+++ test/integration/integration.js
@@ -11,14 +11,16 @@ const execSync = require('child_process').execSync
 const fsExtra = require('fs-extra')
 const Opal = require('../../lib/opal')
 const opalVersion = Opal.get('RUBY_ENGINE_VERSION')
+const exec = require('child_process').exec
 
 RegExp.escape = function(s) {
   return s.replace(/[-\/\\^$*+?.()|[\]{}]/g, '\\$&')
 }
 
 describe('integration', function(){
+  this.timeout(10000)
+
   const tmpDir = path.resolve(__dirname, '../../tmp')
-  const opalCompilerPath = path.resolve(__dirname, '../../vendor/opal-compiler.js')
   const outputBaseDir = path.resolve(tmpDir, 'output')
   const cacheDir = path.join(outputBaseDir, 'cache')
   const fixturesDir = path.resolve(__dirname, '../fixtures')
@@ -57,7 +59,7 @@ describe('integration', function(){
           expect(err).to.be.null
           expect(subject).to.include('Opal.cdecl($scope, \'HELLO\', 123)')
           expect(subject).to.not.match(currentDirectoryExp)
-          expect(runCode()).to.eq('123\n')
+          expect(runCode()).to.eq('123\n\n')
 
           return done()
         })
@@ -71,7 +73,13 @@ describe('integration', function(){
 
   function runCode(otherArgs) {
     const args = otherArgs || ''
-    return execSync(`node -r ${opalCompilerPath} ${args} ${path.join(outputDir, '0.loader.js')} 2>&1 || true`).toString()
+    return execSync(`phantomjs ${path.resolve(__dirname, '../support/runPhantom.js')} ${args} ${path.resolve(outputDir, '0.loader.js')} 2>&1 || true`).toString()
+  }
+
+  // the source-map-support plugin that load_source_maps.js loads makes it easy to test this on node
+  function runSourceMapDependentCode() {
+    const sourceMapFinder = aFixture('load_source_maps.js')
+    return execSync(`node -r ${sourceMapFinder} ${path.join(outputDir, '0.loader.js')} 2>&1 || true`).toString()
   }
 
   beforeEach(function (done) {
@@ -94,6 +102,19 @@ describe('integration', function(){
     assertBasic(config, done)
   })
 
+  // not an end to end test, but since it's a bit slower, put here instead of unit test
+  it('matches our bundler test version', function(done) {
+    this.timeout(5000) // time for shell execute
+
+    const opalVersion = Opal.get('RUBY_ENGINE_VERSION')
+
+    exec('opal -e "puts RUBY_ENGINE_VERSION"', function(err, stdout) {
+      if (err) { done(err) }
+      expect(stdout.trim()).to.eq(opalVersion)
+      return done()
+    })
+  })
+
   it('loads requires', function (done){
     const config = assign({}, globalConfig, {
       entry: aFixture('entry_another_dep.js')
@@ -112,7 +133,7 @@ describe('integration', function(){
           expect(subject).to.include('Opal.cdecl($scope, \'HELLO\', 123)')
           expect(subject).to.include('Opal.cdecl($scope, \'INSIDE\', 789)')
           expect(subject).to.not.match(currentDirectoryExp)
-          expect(runCode()).to.eq('123\nwe made it\n')
+          expect(runCode()).to.eq('123\n\nwe made it\n\n')
 
           return done()
         })
@@ -138,7 +159,7 @@ describe('integration', function(){
           expect(subject).to.include('Opal.cdecl($scope, \'HELLO\', 123)')
           expect(subject).to.include('Opal.cdecl($scope, \'INSIDE\', 789)')
           expect(subject).to.not.match(currentDirectoryExp)
-          expect(runCode()).to.eq('123\nwe made it\n')
+          expect(runCode()).to.eq('123\n\nwe made it\n\n')
 
           return done()
         })
@@ -167,13 +188,30 @@ describe('integration', function(){
           expect(subject).to.not.include('Opal.cdecl($scope, \'HELLO\', 123)')
           expect(subject).to.include('Opal.cdecl($scope, \'INSIDE\', 789)')
           expect(subject).to.not.match(currentDirectoryExp)
-          expect(runCode()).to.eq('we made it\n')
+          expect(runCode()).to.eq('we made it\n\n')
 
           return done()
         })
       })
     })
   })
+
+  // can change the build process to build 2 files, a regular opal build and
+  // then the compiler in a separate file
+  it('the bundled opal version does not include compilation in the webpack bundle')
+
+  // should behave like stubbing opal, opal/mini, opal/full, opal/base
+  // separate from compilation
+  it('allows stubbing Opal requires so they can be provided outside webpack')
+
+// should also stub opal but sub in a configured value instead of vendor/opal-compiler
+    // also rename opal-compiler.js to bundled-opal.js
+    // should behave like stubbing opal, opal/mini, opal/full, opal/base
+  it('allows using a statically provided Opal distro')
+
+// should add Opal to the OPAL_LOAD_PATHS environment variable and
+    // automatically use it
+  it('allows using a bundler provided Opal distro')
 
   it('allows stub inside require', function(done) {
     const config = assign({}, globalConfig, {
@@ -200,7 +238,7 @@ describe('integration', function(){
           expect(subject).to.include('Opal.modules["dependency"]')
           expect(subject).to.include('Opal.modules["another_dependency"]')
           expect(subject).to.include('Opal.modules["inside_load_path"]')
-          expect(runCode()).to.eq('we made it\n')
+          expect(runCode()).to.eq('we made it\n\n')
 
           return done()
         })
@@ -285,7 +323,7 @@ describe('integration', function(){
           // don't want paths hard coded to machines in here
           expect(subject).to.not.match(currentDirectoryExp)
           expect(subject).to.include('Opal.modules["tree/file1"]')
-          expect(runCode()).to.eq('inside the tree\n')
+          expect(runCode()).to.eq('inside the tree\n\n')
 
           return done()
         })
@@ -294,8 +332,6 @@ describe('integration', function(){
   })
 
   it('outputs correct source maps', function (done) {
-    this.timeout(10000)
-
     const config = assign({}, globalConfig, {
       entry: aFixture('entry_source_maps.js'),
       devtool: 'source-map'
@@ -307,7 +343,7 @@ describe('integration', function(){
       fs.readdir(outputDir, (err, files) => {
         expect(err).to.be.null
         expect(files).to.have.length(2)
-        var output = runCode('-r '+aFixture('load_source_maps.js'))
+        var output = runSourceMapDependentCode()
         // ruby output, might need some more work since we're 1 line off
         // expecting test/fixtures/source_maps.rb:4:in `hello': source map location (RuntimeError)
         expect(output).to.match(/output\/loader\/webpack:\/test\/fixtures\/source_maps\.rb:3:1\)/)
@@ -318,8 +354,6 @@ describe('integration', function(){
   })
 
   it('outputs correct source maps when stubs are used', function (done) {
-    this.timeout(10000)
-
     const config = assign({}, globalConfig, {
       entry: aFixture('entry_source_maps_stubs.js'),
       devtool: 'source-map',
@@ -334,11 +368,11 @@ describe('integration', function(){
       fs.readdir(outputDir, (err, files) => {
         expect(err).to.be.null
         expect(files).to.have.length(2)
-        var output = runCode('-r '+aFixture('load_source_maps.js'))
+        var output = runSourceMapDependentCode()
         // ruby output, might need some more work since we're 1 line off
         // expecting test/fixtures/source_maps.rb:4:in `hello': source map location (RuntimeError)
-        expect(output).to.match(/output\/loader\/webpack:\/test\/fixtures\/source_maps_stubs\.rb:6:1\)/)
-        expect(output).to.match(/output\/loader\/webpack:\/test\/fixtures\/source_maps_stubs.rb:10:1\)/)
+        expect(output).to.match(/output\/loader\/webpack:\/test\/fixtures\/source_maps_stubs\.rb:7:1\)/)
+        expect(output).to.match(/output\/loader\/webpack:\/test\/fixtures\/source_maps_stubs.rb:11:1\)/)
         return done()
       })
     })
@@ -371,7 +405,7 @@ describe('integration', function(){
 
           expect(err).to.be.null
           expect(subject).to.not.match(currentDirectoryExp)
-          expect(runCode()).to.eq('[Object#onearg] wrong number of arguments(0 for 1)\n[Object#two_arg] wrong number of arguments(1 for 2)\n')
+          expect(runCode()).to.eq('[Object#onearg] wrong number of arguments(0 for 1)\n\n[Object#two_arg] wrong number of arguments(1 for 2)\n\n')
 
           return done()
         })
@@ -393,7 +427,8 @@ describe('integration', function(){
       }
       else {
         // Opal 0.10 regression - https://github.com/opal/opal/pull/1426
-        expect(error.message).to.match(/Module build failed.*false/)
+        // and https://github.com/opal/opal/issues/1427
+        expect(error.message).to.match(/Module build failed.*An error occurred while compiling:.*error/)
       }
       return done()
     })
@@ -419,7 +454,8 @@ describe('integration', function(){
       assertBasic(config, () => {
         fs.readdir(cacheDir, (err, files) => {
           expect(err).to.be.null
-          expect(files).to.have.length(1)
+          // 1 file + opal
+          expect(files).to.have.length(2)
           return done()
         })
       })
@@ -447,7 +483,8 @@ describe('integration', function(){
 
       fs.readdir(cacheDir, (err, files) => {
         expect(err).to.be.null
-        expect(files).to.have.length(3)
+        // 3 dependencies + opal
+        expect(files).to.have.length(4)
         return done()
       })
     })
@@ -476,7 +513,7 @@ describe('integration', function(){
       fs.readdir(outputDir, (err, files) => {
         expect(err).to.be.null
         expect(files).to.have.length(2)
-        var output = runCode('-r '+aFixture('load_source_maps.js'))
+        var output = runSourceMapDependentCode()
         // ruby output, might need some more work since we're 1 line off
         // expecting test/fixtures/source_maps.rb:4:in `hello': source map location (RuntimeError)
         expect(output).to.match(/output\/loader\/webpack:\/test\/fixtures\/source_maps\.rb:3:1\)/)