+++ test/unit/getCompiler.js
@@ -1,6 +1,11 @@
 'use strict'
 
 const expect = require('chai').expect
+const path = require('path')
+const tmp = require('tmp')
+const fs = require('fs')
+const exec = require('child_process').exec
+
 const getCompiler = require('../../lib/getCompiler')
 
 describe('compiler', function(){
@@ -14,10 +19,71 @@ describe('compiler', function(){
     return compiler.$result()
   }
 
+  it('loads an Opal compiler from a confgurable file', function(done) {
+    const code = `const getCompiler = require('./lib/opal')\nconsole.log(Opal.get('RUBY_ENGINE_VERSION'))`
+
+    tmp.file(function (err, tmpPath, fd, cleanup) {
+      fs.writeFile(tmpPath, code, function (err) {
+        if (err) {
+          cleanup()
+          done(err)
+        }
+
+        exec(`node ${tmpPath}`, {
+          env: {
+            OPAL_COMPILER_PATH: path.resolve(__dirname, '../support/tweakedOpalCompiler.js')
+          }
+        },
+        function (err, stdout) {
+          if (err) {
+            cleanup()
+            done(err)
+          }
+
+          try {
+            expect(stdout.trim()).to.eq('0.10.0.beta2.webpacktest')
+            return done()
+          }
+          finally {
+            cleanup()
+          }
+        })
+      })
+    })
+  })
+
+  // will need to call bundler and deal with this on the fly
+  // since we don't want the compiler in the bundled code
+  // we'll probably want to use the same process we do for the build
+  // and stick the "compiled" file somewhere as a cache
+  // could reuse that between the build process and runtime
+  it('can fetch an Opal compiler from Bundler', function() {
+  })
+
   it('raw compiler works', function(){
     var result = doCompile('foo', 'puts "Howdy #{1+2}"')
 
     expect(result).to.include('self.$puts("Howdy " + ($rb_plus(1, 2)))')
+  })
+
+  it('handles syntax errors', function(done) {
+    // was having problems with chai expect throw assertions
+    let error = null
+    try {
+      doCompile('foo', 'def problem')
+    }
+    catch (e) {
+      error = e
+    }
+    if (error) {
+      if (error.name === 'SyntaxError' && /An error occurred while compiling: foo[\S\s]*false/.test(error.message)) {
+        return done()
+      }
+      else {
+        return done(new Error(`Unexpected error ${error}`))
+      }
+    }
+    return done(new Error('expected error, got none'))
   })
 
   it('passes on compiler options', function() {