+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 MASTER:
 
+* By default, require 'opal' includes bundled opal in this package, you are discouraged from mixing this compiler with your own runtime
 * File caching between runs
 * Supports passing stubs in via global options
 * Avoid absolute path in opal module names (e.g. `Opal.modules['/root/something']`)